MyStatsPackage.rse_sum — Function

rse_sum(V::Vector)::Real

Return the sum of all elements included in a vector .

Arguments

V – Vector of floats/ints

Output

A real number representing the sum of all the vectro elements

Notes

Make sure not to include NaNs or character vectors

Examples

julia> rse_mean([1,5,6])
12

julia> rse_mean(1:36)
666

See also rsemean, rsestd, rse_tstat